feat: adds ability to filter journey search history
